+++ src/TextTable.php
@@ -35,8 +35,6 @@ use function implode;
 use function is_integer;
 use function is_null;
 use function is_string;
-use function str_pad;
-use function strlen;
 use function substr;
 use const false;
 use const null;
@@ -50,7 +48,7 @@ use const true;
  *
  * @package Inane\Cli
  *
- * @version 0.2.1
+ * @version 0.2.2
  */
 class TextTable implements Stringable {
     /**
@@ -214,9 +212,9 @@ class TextTable implements Stringable {
         $sd = $this->config->column->definition;
 
         for ($i = 0; $i < count($row); $i++)
-            if (count($ad) < ($i + 1) || strlen($row[$i]) > $ad[$i]) {
-                if ($this->getDefinitionRule() == DefinitionRule::Max && strlen($row[$i]) > $sd[$i]) $ad[$i] = $sd[$i];
-                else $ad[$i] = strlen($row[$i]);
+            if (count($ad) < ($i + 1) || Colors::width($row[$i]) > $ad[$i]) {
+                if ($this->getDefinitionRule() == DefinitionRule::Max && Colors::width($row[$i]) > $sd[$i]) $ad[$i] = $sd[$i];
+                else $ad[$i] = Colors::width($row[$i]);
             }
 
         return count($this->getColumnDefinition()) == count($row);
@@ -333,8 +331,8 @@ class TextTable implements Stringable {
                 $r = $this->getDivider();
 
             for ($i = 0; $i < count($this->getColumnDefinition()); $i++) {
-                $col = str_pad($r[$i], $this->getColumnDefinition()[$i]);
-                if ($this->getDefinitionRule()->truncate() && strlen($col) > $this->getColumnDefinition()[$i]) $col = substr($col, 0, $this->getColumnDefinition()[$i] - 1) . '>';
+                $col = Colors::pad($r[$i], $this->getColumnDefinition()[$i]);
+                if ($this->getDefinitionRule()->truncate() && Colors::width($col) > $this->getColumnDefinition()[$i]) $col = substr($col, 0, $this->getColumnDefinition()[$i] - 1) . '>';
                 $cols[] = $col;
             }