+++ CHANGELOG.md
@@ -6,4 +6,6 @@
 
 ### current: 0.12.1 @2022 Aug xx
 
- - fix: TextTable divider => uses full column width
+ - fix: `TextTable` `divider` => uses full column width
+ - fix: `TextTable` coloured string cell width
+ - fix: `Colors::length` error